Prepare the BoardGame page. Add the needed Model that will be used to populate the page. Rename variables for consitency.
