+++ Boardium/Boardium/Views/Games/BoardGame.cshtml
@@ -1,0 +1,65 @@
+﻿@model Boardium.Models.BoardGameViewModel;
+@{
+    ViewData["Title"] = Model.Title;
+}
+
+<div class="container mt-5">
+    <h1 class="mb-4">@Model.Title</h1>
+
+    <!-- Image Carousel -->
+    <div id="gameCarousel" class="carousel slide mb-4" data-bs-ride="carousel">
+        <div class="carousel-inner">
+            @for (int i = 0; i < Model.PathsToImages.Count(); i++)
+            {
+                <div class="carousel-item @(i == 0 ? "active" : "")">
+                    <img src="@Model.PathsToImages[i]" class="d-block w-100" alt="Game image @i">
+                </div>
+            }
+        </div>
+        <button class="carousel-control-prev" type="button" data-bs-target="#gameCarousel" data-bs-slide="prev">
+            <span class="carousel-control-prev-icon"></span>
+        </button>
+        <button class="carousel-control-next" type="button" data-bs-target="#gameCarousel" data-bs-slide="next">
+            <span class="carousel-control-next-icon"></span>
+        </button>
+    </div>
+
+    <!-- Game Info -->
+    <div class="row">
+        <div class="col-md-8">
+            <h4>Description</h4>
+            <p>@Model.Description</p>
+
+            <ul class="list-group list-group-flush">
+                <li class="list-group-item"><strong>Players:</strong> @Model.MinPlayers - @Model.MaxPlayers</li>
+                <li class="list-group-item"><strong>Playing Time:</strong> @Model.PlayingTimeMinutes minutes</li>
+                <li class="list-group-item"><strong>Categories:</strong>
+                    @foreach (var category in Model.Categories)
+                    {
+                        <span class="badge bg-secondary me-1">@category</span>
+                    }
+                </li>
+            </ul>
+
+            <h4 class="mt-4">Publisher</h4>
+            <p><strong>@Model.publisher.Name</strong><br />
+               @Model.publisher.Website</p>
+        </div>
+
+        <!-- Copies -->
+        <div class="col-md-4">
+            <h4>Game Copies</h4>
+            <!--<ul class="list-group">-->
+               <!--< @@foreach (var copy in Model.Copies) -->
+                <!--{-->
+                    <!--<li class="list-group-item d-flex justify-content-between align-items-center">-->
+                        <!--Copy #@@copy.Id-->
+                        <!--<span class="badge @@(copy.IsAvailable ? "bg-success" : "bg-danger")">-->
+                           <!-- @@(copy.IsAvailable ? "Available" : "Unavailable")-->
+                        <!--</span>-->
+                    <!--</li>-->
+                <!--}-->
+            <!--</ul>-->
+        </div>
+    </div>
+</div>
