--- Boardium/Boardium/Views/Games/BoardGames.cshtml
@@ -1,6 +1,0 @@
-﻿@{
-    ViewData["Title"] = "Board Game";
-}
-
-<h1>Index</h1>
-
